split app.py into multiple files
