+++ src/data/load_data.py
@@ -1,0 +1,4 @@
+import pandas as pd
+
+# Load processed salaries data
+data = pd.read_csv('data/processed/salaries.csv')
